+++ routes/courses.js
@@ -1,7 +1,9 @@
 const express = require('express');
+const base_url = require('../baseURLs/courses')
 const { allCourses } = require('../controllers/courses');
 const router = express.Router({ mergeParams: true });
 
-router.route(`/`).get(allCourses);
+
+router.route(base_url).get(allCourses);
 
 module.exports = router;
