+++ server.js
@@ -43,6 +43,6 @@ const server = app.listen(PORT, () => {
 
 //Handle unhandled promise rejections
 process.on('unhandledRejection', (err) => {
-  console.log(`Error: ${err}`);
+  console.log(`Error: ${err}`).red;
   server.close(() => process.exit(1));
 });
